Refactored the building of Solr documents into its own component. Added Javadocs.
