+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/search/service/solr/indexer/CatalogDocumentBuilderImpl.java
@@ -1,0 +1,270 @@
+/*
+ * #%L
+ * BroadleafCommerce Framework
+ * %%
+ * Copyright (C) 2009 - 2019 Broadleaf Commerce
+ * %%
+ * Licensed under the Broadleaf Fair Use License Agreement, Version 1.0
+ * (the "Fair Use License" located  at http://license.broadleafcommerce.org/fair_use_license-1.0.txt)
+ * unless the restrictions on use therein are violated and require payment to Broadleaf in which case
+ * the Broadleaf End User License Agreement (EULA), Version 1.1
+ * (the "Commercial License" located at http://license.broadleafcommerce.org/commercial_license-1.1.txt)
+ * shall apply.
+ * 
+ * Alternatively, the Commercial License may be replaced with a mutually agreed upon license (the "Custom License")
+ * between you and Broadleaf Commerce. You may not use this file except in compliance with the applicable license.
+ * #L%
+ */
+package org.broadleafcommerce.core.search.service.solr.indexer;
+
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.apache.solr.common.SolrInputDocument;
+import org.broadleafcommerce.common.exception.ExceptionHelper;
+import org.broadleafcommerce.common.extension.ExtensionResultStatusType;
+import org.broadleafcommerce.common.locale.domain.Locale;
+import org.broadleafcommerce.common.sandbox.SandBoxHelper;
+import org.broadleafcommerce.common.util.GenericOperation;
+import org.broadleafcommerce.common.util.HibernateUtils;
+import org.broadleafcommerce.core.catalog.domain.Indexable;
+import org.broadleafcommerce.core.catalog.domain.Product;
+import org.broadleafcommerce.core.search.dao.CatalogStructure;
+import org.broadleafcommerce.core.search.domain.Field;
+import org.broadleafcommerce.core.search.domain.IndexField;
+import org.broadleafcommerce.core.search.domain.IndexFieldType;
+import org.broadleafcommerce.core.search.domain.solr.FieldType;
+import org.broadleafcommerce.core.search.service.solr.SolrConfiguration;
+import org.broadleafcommerce.core.search.service.solr.SolrHelperService;
+import org.broadleafcommerce.core.search.service.solr.index.SolrIndexCachedOperation;
+import org.broadleafcommerce.core.search.service.solr.index.SolrIndexServiceExtensionManager;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Component;
+
+import java.lang.reflect.InvocationTargetException;
+import java.math.BigDecimal;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+
+import javax.annotation.Resource;
+
+/**
+ * Default implementation of the {@link DocumentBuilder} to build {@link SolrInputDocument}s from {@link Product}s.
+ * 
+ * @author Kelly Tisdell
+ *
+ */
+@Component("blCatalogDocumentBuilder")
+public class CatalogDocumentBuilderImpl implements CatalogDocumentBuilder {
+    
+    private static final Log LOG = LogFactory.getLog(CatalogDocumentBuilderImpl.class);
+    
+    @Resource(name = "blSolrHelperService")
+    protected SolrHelperService shs;
+    
+    @Resource(name = "blSolrIndexServiceExtensionManager")
+    protected SolrIndexServiceExtensionManager extensionManager;
+    
+    @Qualifier("blCatalogSolrConfiguration")
+    @Autowired(required = false)
+    protected SolrConfiguration solrConfiguration;
+    
+    @Resource(name = "blSandBoxHelper")
+    protected SandBoxHelper sandBoxHelper;
+
+    @Override
+    public SolrInputDocument buildDocument(final Indexable indexable, final List<IndexField> fields, final List<Locale> locales) {
+        try {
+            final SolrInputDocument document = new SolrInputDocument();
+            HibernateUtils.executeWithoutCache(new GenericOperation<Void>() {
+                @Override
+                public Void execute() throws Exception {
+                    attachBasicDocumentFields(indexable, document);
+                    attachIndexableDocumentFields(document, indexable, fields, locales);
+                    attachAdditionalDocumentFields(indexable, document);
+                    extensionManager.getProxy().attachChildDocuments(indexable, document, fields, locales);
+                    return null;
+                }
+            });
+            modifyDocument(indexable, fields, locales);
+            return document;
+        } catch (Exception e) {
+            LOG.warn("An error occured trying to build a SolrInputDocument for Indexable of type, " + indexable.getClass().getName() + " with an ID of " + indexable.getId(), e);
+            return null;
+        }
+    }
+    
+    protected void attachIndexableDocumentFields(SolrInputDocument document, Indexable indexable, List<IndexField> fields, List<Locale> locales) {
+        for (IndexField indexField : fields) {
+            try {
+                // If we find an IndexField entry for this field, then we need to store it in the index
+                if (indexField != null) {
+                    List<IndexFieldType> searchableFieldTypes = indexField.getFieldTypes();
+
+                    // For each of its search field types, get the property values, and add a field to the document for each property value
+                    for (IndexFieldType sft : searchableFieldTypes) {
+                        FieldType fieldType = sft.getFieldType();
+                        Map<String, Object> propertyValues = getPropertyValues(indexable, indexField.getField(), fieldType, locales);
+
+                        ExtensionResultStatusType result = extensionManager.getProxy().populateDocumentForIndexField(document, indexField, fieldType, propertyValues);
+
+                        if (ExtensionResultStatusType.NOT_HANDLED.equals(result)) {
+                            // Build out the field for every prefix
+                            for (Entry<String, Object> entry : propertyValues.entrySet()) {
+                                String prefix = entry.getKey();
+                                prefix = StringUtils.isBlank(prefix) ? prefix : prefix + "_";
+
+                                String solrPropertyName = shs.getPropertyNameForIndexField(indexField, fieldType, prefix);
+                                Object value = entry.getValue();
+
+                                if (FieldType.isMultiValued(fieldType)) {
+                                    document.addField(solrPropertyName, value);
+                                } else {
+                                    document.setField(solrPropertyName, value);
+                                }
+                            }
+                        }
+                    }
+                }
+
+            } catch (Exception e) {
+                LOG.error("Could not get value for property[" + indexField.getField().getQualifiedFieldName() + "] for product id["
+                        + indexable.getId() + "]", e);
+                throw ExceptionHelper.refineException(e);
+            }
+        }
+    }
+    
+    protected void attachAdditionalDocumentFields(Indexable indexable, SolrInputDocument document) {
+        //Empty implementation. Placeholder for others to extend and add additional fields
+        extensionManager.getProxy().attachAdditionalDocumentFields(indexable, document);
+    }
+    
+    protected void attachBasicDocumentFields(Indexable indexable, SolrInputDocument document) {
+        CatalogStructure cache = SolrIndexCachedOperation.getCache();
+        if (cache == null) {
+            String msg = "SolrIndexService.performCachedOperation() must be used in conjuction with"
+                + " solrIndexDao.populateProductCatalogStructure() in order to correctly build catalog documents or should"
+                + " be invoked from buildIncrementalIndex()";
+            throw new IllegalStateException(msg);
+        }
+
+        // Add the namespace and ID fields for this product
+        document.addField(shs.getNamespaceFieldName(), solrConfiguration.getNamespace());
+        document.addField(shs.getIdFieldName(), shs.getSolrDocumentId(document, indexable));
+        document.addField(shs.getTypeFieldName(), shs.getDocumentType(indexable));
+        document.addField(shs.getIndexableIdFieldName(), shs.getIndexableId(indexable));
+
+        extensionManager.getProxy().attachAdditionalBasicFields(indexable, document, shs);
+
+        Long cacheKey = this.shs.getCurrentProductId(indexable); // current
+        if (!cache.getParentCategoriesByProduct().containsKey(cacheKey)) {
+            cacheKey = sandBoxHelper.getOriginalId(cacheKey); // parent
+            if (!cache.getParentCategoriesByProduct().containsKey(cacheKey)) {
+                cacheKey = shs.getIndexableId(indexable); // master
+            }
+        }
+
+        // TODO: figure this out more generally; this doesn't work for CMS content
+        // The explicit categories are the ones defined by the product itself
+        if (cache.getParentCategoriesByProduct().containsKey(cacheKey)) {
+            for (Long categoryId : cache.getParentCategoriesByProduct().get(cacheKey)) {
+                document.addField(shs.getExplicitCategoryFieldName(), shs.getCategoryId(categoryId));
+
+                // Make sure that we're always referencing the parent for the sort field
+                String categorySortFieldName = shs.getCategorySortFieldName(shs.getCategoryId(categoryId));
+                // The issue here was the super category id is always what is stored in the cache, while the category
+                // by product id is the overridden versions. Need to always look at parent version for cache stuff, which
+                // is given from shs.getCategoryId
+                // First try the current level
+                String displayOrderKey = categoryId + "-" + cacheKey;
+                Long displayOrder = convertDisplayOrderToLong(cache, displayOrderKey);
+                if (displayOrder == null) {
+                    // Didn't find the cache at the current level, this might be an override so look upwards
+                    displayOrderKey = shs.getCategoryId(categoryId) + "-" + cacheKey;
+                    displayOrder = convertDisplayOrderToLong(cache, displayOrderKey);
+                }
+                
+                if (document.getField(categorySortFieldName) == null && displayOrder != null) {
+                    document.addField(categorySortFieldName, displayOrder);
+                }
+
+                // This is the entire tree of every category defined on the product
+                buildFullCategoryHierarchy(document, cache, categoryId, new HashSet<Long>());
+            }
+        }
+    }
+    
+    protected void buildFullCategoryHierarchy(SolrInputDocument document, CatalogStructure cache, Long categoryId, Set<Long> indexedParents) {
+        Long catIdToAdd = shs.getCategoryId(categoryId); 
+
+        Collection<Object> existingValues = document.getFieldValues(shs.getCategoryFieldName());
+        if (existingValues == null || !existingValues.contains(catIdToAdd)) {
+            document.addField(shs.getCategoryFieldName(), catIdToAdd);
+        }
+
+        Set<Long> parents = cache.getParentCategoriesByCategory().get(categoryId);
+        for (Long parent : parents) {
+            if (!indexedParents.contains(parent)) {
+                indexedParents.add(parent);
+                buildFullCategoryHierarchy(document, cache, parent, indexedParents);
+            }
+        }
+    }
+    
+    protected Map<String, Object> getPropertyValues(Indexable indexedItem, Field field, FieldType fieldType, List<Locale> locales)
+            throws IllegalAccessException, InvocationTargetException, NoSuchMethodException {
+
+        String propertyName = field.getPropertyName();
+        Map<String, Object> values = new HashMap<>();
+
+        ExtensionResultStatusType extensionResult = ExtensionResultStatusType.NOT_HANDLED;
+        if (extensionManager != null) {
+            extensionResult = extensionManager.getProxy().addPropertyValues(indexedItem, field, fieldType, values, propertyName, locales);
+        }
+        
+        if (ExtensionResultStatusType.NOT_HANDLED.equals(extensionResult)) {
+            Object propertyValue = shs.getPropertyValue(indexedItem, field);
+            if (propertyValue != null) {
+                values.put("", propertyValue);
+            }
+        }
+
+        return values;
+    }
+    
+    /**
+     *  We multiply the BigDecimal by 1,000,000 to maintain any possible decimals in use the
+     *  displayOrder value.
+     *
+     * @param cache
+     * @param displayOrderKey
+     * @return
+     */
+    protected Long convertDisplayOrderToLong(CatalogStructure cache, String displayOrderKey) {
+        BigDecimal displayOrder = cache.getDisplayOrdersByCategoryProduct().get(displayOrderKey);
+
+        if (displayOrder == null) {
+            return null;
+        }
+
+        return displayOrder.multiply(BigDecimal.valueOf(1000000)).longValue();
+    }
+    
+    /**
+     * This is a simple hook point to allow implementors to override and modify the documents after the default functionality has created them.
+     * @param indexable
+     * @param fields
+     * @param locales
+     */
+    protected void modifyDocument(final Indexable indexable, final List<IndexField> fields, final List<Locale> locales) {
+        //Do nothing by default.
+    }
+    
+}
