+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/search/service/solr/indexer/DocumentBuilder.java
@@ -1,0 +1,37 @@
+/*
+ * #%L
+ * BroadleafCommerce Framework
+ * %%
+ * Copyright (C) 2009 - 2019 Broadleaf Commerce
+ * %%
+ * Licensed under the Broadleaf Fair Use License Agreement, Version 1.0
+ * (the "Fair Use License" located  at http://license.broadleafcommerce.org/fair_use_license-1.0.txt)
+ * unless the restrictions on use therein are violated and require payment to Broadleaf in which case
+ * the Broadleaf End User License Agreement (EULA), Version 1.1
+ * (the "Commercial License" located at http://license.broadleafcommerce.org/commercial_license-1.1.txt)
+ * shall apply.
+ * 
+ * Alternatively, the Commercial License may be replaced with a mutually agreed upon license (the "Custom License")
+ * between you and Broadleaf Commerce. You may not use this file except in compliance with the applicable license.
+ * #L%
+ */
+package org.broadleafcommerce.core.search.service.solr.indexer;
+
+import org.apache.solr.common.SolrInputDocument;
+import org.broadleafcommerce.common.locale.domain.Locale;
+import org.broadleafcommerce.core.catalog.domain.Indexable;
+import org.broadleafcommerce.core.search.domain.IndexField;
+
+import java.util.List;
+
+/**
+ * Component to build a Solr document from an {@link Indexable}.
+ * 
+ * @author Kelly Tisdell
+ *
+ */
+public interface DocumentBuilder {
+
+    public SolrInputDocument buildDocument(Indexable indexable, List<IndexField> fields, List<Locale> locales);
+    
+}
